==> processing/2.binding_functional.ipynb <==
# B & F Assays

from pandas import read_csv, isna

from assay_processing import delete_assays
from assay_fusion import cluster_assays, merge_assays

file_path = 'deliverables/chembl_extract.csv'
types_to_keep = ['IC50', 'EC50', 'Ki', 'Kd', 'Potency', 'AC50']
min_values_per_assay = 50

df = read_csv(file_path, low_memory=False)
df = df.loc[df['assay_type'].isin(['B', 'F'])]
df = df.loc[df['standard_type'].isin(types_to_keep)]

---
## only keep assays with at least one pchembl value

assays = df.assay_id.unique()

assay_ids_to_drop = list()
for assay in assays:
    unique_pchembl_values = df[df.assay_id==assay]['pchembl_value'].unique()
    if len(unique_pchembl_values) == 1 and isna(unique_pchembl_values[0]):
        assay_ids_to_drop.append(assay)        

print(f'Before: {len(assays)}')

print(f'Bad: {len(assay_ids_to_drop)}')

df = delete_assays(assay_ids_to_drop, df)

### 2997 remaining assays

---
## Statistical Assay Fusion

%%time
assay_groups = cluster_assays(df)

from pickle import dump
reference = list()
for group in assay_groups:
    m = min(group)
    reference.extend([{'assay_id': a, 'input_assay_id': m} for a in group])
with open('reference_bf.pkl', 'wb') as f:
    dump(reference, f)

df = merge_assays(df, assay_groups)

df.assay_id.unique().shape[0]

---
## Only keep assays with at least 50 data points

assays = df.assay_id.unique()

assay_ids_to_drop = list()
for assay in assays:
    assay_values = [v for v in df[df.assay_id==assay]['standard_value'] if not isna(v)] 
    if len(assay_values) < min_values_per_assay:
        assay_ids_to_drop.append(assay)   

print(f'Before: {len(assays)}')

print(f'Bad: {len(assay_ids_to_drop)}')

df = delete_assays(assay_ids_to_drop, df)

### 2979 remaining assays

---
## Unit filtering

df = df[df.standard_units == 'nM']

---
## Export

df.to_csv('data/binding_functional.csv', index=False)
