==> processing/4.binning_export.ipynb <==
from pandas import read_csv, concat

from assay_processing import (
    log_to_nanomol, 
    bin_assays, 
    delete_assays, 
    build_t2, 
    build_t3, 
    build_t4
)

---
## Read Data

solubility_df = read_csv('data/physicochemical_solubility.csv')
logP_df = read_csv('data/physicochemical_logp.csv')
logD_df = read_csv('data/physicochemical_logd.csv')

---
## Reset Assay IDs

reference = list()

start = 2000000
for i, label in enumerate(logD_df.pH.unique()):
    for a in logD_df.loc[logD_df['pH'] == label, 'assay_id']:
        reference.append({'assay_id': a, 'input_assay_id': start + i})
    logD_df.loc[logD_df['pH'] == label, 'assay_id'] = start + i
    
start = 3000000 
for i, label in enumerate(solubility_df.pH_solvent.unique()):
    for a in solubility_df.loc[solubility_df['pH_solvent'] == label, 'assay_id']:
        reference.append({'assay_id': a, 'input_assay_id': start + i})
    solubility_df.loc[solubility_df['pH_solvent'] == label, 'assay_id'] = start + i
    
for a in logP_df.assay_id.unique():
    reference.append({'assay_id': a, 'input_assay_id': 4000000})
logP_df.assay_id = 4000000

from pickle import dump
with open('reference_pc.pkl', 'wb') as f:
    dump(reference, f)

---
## All standard values must be in nM

solubility_df['standard_value'] = solubility_df.logS.apply(log_to_nanomol)
logD_df['standard_value'] = logD_df.standard_value.apply(log_to_nanomol)
logP_df['standard_value'] = logP_df.standard_value.apply(log_to_nanomol)

columns = set(solubility_df.columns)
columns.intersection_update(logD_df.columns)
columns.intersection_update(logP_df.columns)

physico_chemical_df = concat([solubility_df[columns], logD_df[columns], logP_df[columns]])

physico_chemical_df.to_csv('data/physico_chemical.csv', index=False)

---

input_directory = 'data'
output_directory = 'deliverables'

files = [
    'admet_tox.csv',
    'binding_functional.csv',
    'physico_chemical.csv',
]
file_paths = [input_directory + '/' + f for f in files]

dfs = [read_csv(file_path, low_memory=False) for file_path in file_paths]
df_ = concat(dfs, sort=False)

t3_data, t4_data, assay_ids_to_drop = bin_assays(df_)
df = delete_assays(assay_ids_to_drop, df_)

t2_df = build_t2(df)
t3_df = build_t3(t3_data)
t4_df = build_t4(t4_data)

t2_df.head(10)

t3_df.head(10)

t4_df.head(10)

t2_df.to_csv(output_directory + '/T2.csv', index=False)
t3_df.to_csv(output_directory + '/T3.csv', index=False)
t4_df.to_csv(output_directory + '/T4.csv', index=False)
