==> processing/5.analysis.ipynb <==
from assay_processing import *

from pandas import read_csv

---
# P

solubility_df = read_csv('data/physicochemical_solubility.csv')
logP_df = read_csv('data/physicochemical_logp.csv')
logD_df = read_csv('data/physicochemical_logd.csv')

start = 2000000
for i, label in enumerate(logD_df.pH.unique()):
    logD_df.loc[logD_df['pH'] == label, 'assay_id'] = start + i
    
start = 3000000 
for i, label in enumerate(solubility_df.pH_solvent.unique()):
    solubility_df.loc[solubility_df['pH_solvent'] == label, 'assay_id'] = start + i
    
logP_df.assay_id = 4000000

soluability_df['standard_value'] = solubility_df.logS.apply(log_to_nanomol)
logD_df['standard_value'] = logD_df.standard_value.apply(log_to_nanomol)
logP_df['standard_value'] = logP_df.standard_value.apply(log_to_nanomol)

print(f'Solubility assays: {len(solubility_df.assay_id.unique())}')
print(f'Log D assays: {len(logD_df.assay_id.unique())}')
print(f'Log P assays: {len(logP_df.assay_id.unique())}')

t3_data, t4_data, assay_ids_to_drop = bin_assays(solubility_df)
solubility_df = delete_assays(assay_ids_to_drop, solubility_df)
t3_solubility_df = build_t3(t3_data)
t4_solubility_df = build_t4(t4_data)

t3_data, t4_data, assay_ids_to_drop = bin_assays(logD_df)
logD_df = delete_assays(assay_ids_to_drop, logD_df)
t3_logD_df = build_t3(t3_data)
t4_logD_df = build_t4(t4_data)

t3_data, t4_data, assay_ids_to_drop = bin_assays(logP_df)
logP_df = delete_assays(assay_ids_to_drop, logP_df)
t3_logP_df = build_t3(t3_data)
t4_logP_df = build_t4(t4_data)

print(f'Solubility assays: {len(solubility_df.assay_id.unique())}')
print(f'Log D assays: {len(logD_df.assay_id.unique())}')
print(f'Log P assays: {len(logP_df.assay_id.unique())}')

print(f'Solubility tasks: {len(t3_solubility_df.classification_task_id.unique())}')
print(f'Log D tasks: {len(t3_logD_df.classification_task_id.unique())}')
print(f'Log P tasks: {len(t3_logP_df.classification_task_id.unique())}')

---
# A&T

at_df = read_csv('data/admet_tox.csv', low_memory=False)

print(f'A&T assays: {len(at_df.assay_id.unique())}')

t3_data, t4_data, assay_ids_to_drop = bin_assays(at_df)
at_df = delete_assays(assay_ids_to_drop, at_df)
t3_at_df = build_t3(t3_data)
t4_at_df = build_t4(t4_data)

print(f'A&T assays: {len(at_df.assay_id.unique())}')
print(f'A&T tasks: {len(t3_at_df.classification_task_id.unique())}')

---
# B&F

bf_df = read_csv('data/binding_functional.csv', low_memory=False)

print(f'B&F assays: {len(bf_df.assay_id.unique())}')

t3_data, t4_data, assay_ids_to_drop = bin_assays(bf_df)
bf_df = delete_assays(assay_ids_to_drop, bf_df)
t3_bf_df = build_t3(t3_data)
t4_bf_df = build_t4(t4_data)

print(f'B&F assays: {len(bf_df.assay_id.unique())}')
print(f'B&F tasks: {len(t3_bf_df.classification_task_id.unique())}')

--- 
# All

t3_global_df = read_csv('deliverables/T3.csv', low_memory=False)

print(f'All assays: {len(t3_global_df.input_assay_id.unique())}')
print(f'All tasks: {len(t3_global_df.classification_task_id.unique())}')

t3_global_df.input_assay_id.value_counts().value_counts()
