==> processing/6.assay_reference.ipynb <==
from pickle import load

from pandas import DataFrame

files = ['reference_at.pkl', 'reference_bf.pkl', 'reference_pc.pkl']

reference = list()
for file in files:
    with open(file, 'rb') as handle:
        reference.extend(load(handle))

df = DataFrame(reference)

df.to_csv('deliverables/assay_reference.csv', index=False)
