==> processing/8.update_to_year_2_format.ipynb <==
import pandas as pd
import numpy as np

## Loading the data

# Use the relevant folder where you have T2, T3 and T4 
# (can be downloaded from here: https://app.box.com/folder/114032352224 )
data_folder = '../../../public-data-preparation/processing/deliverables/regression'
T2 = pd.read_csv(f'{data_folder}/T2.csv')
T3 = pd.read_csv(f'{data_folder}/T3.csv')
T4 = pd.read_csv(f'{data_folder}/T4.csv')

## Creating new inputs

### T0:

- `use_in_regression` column added and set to True for all assays
- `direction` column added and set to high for all ADME assays
- Five `expert_threshold_{i}` columns added and all set to NaN

T3.head()

T3['use_in_regression'] = True
T3.loc[T3['assay_type'] == 'ADME', 'direction'] = 'high'
for i in range(1, 6):
    T3[f'expert_threshold_{i}'] = np.nan

T0 = T3[['assay_type', 'use_in_regression', 'expert_threshold_1', 'expert_threshold_2',
         'expert_threshold_3', 'expert_threshold_4', 'expert_threshold_5', 'direction']]

print(T0.shape)
print(T0.columns)
T0.head()

### T1

- All values with unit nM converted to log space (pIC/EC50)
- Corresponding qualifiers inverted
- `input_assay_id` can be extracted from a join of T4 on regression_task_id with T3

T4.head()

T4.standard_units.value_counts()

def convert_to_log_scale(x, epsilon=1e-16):
    return -np.log10(1e-9*x + epsilon)

T4.loc[T4['standard_units'] == 'nM', 'standard_value'] = T4.loc[T4['standard_units'] == 'nM', 'standard_value'].apply(convert_to_log_scale)


T4.standard_relation.value_counts()

qualifier_conversion = {
    '>': '<',
    '>=': '<=',
    '<': '>',
    '<=': '>=',
    '<<': '>>',
    '>>': '<<',
    '=': '=',
    '~': '~'
}
T4.loc[T4['standard_units'] == 'nM', 'standard_qualifier'] = T4.loc[T4['standard_units'] == 'nM', 'standard_relation'].map(qualifier_conversion)
T4.loc[T4['standard_units'] != 'nM', 'standard_qualifier'] = T4.loc[T4['standard_units'] != 'nM', 'standard_relation']


T4_join = T4.join(T3.set_index('regression_task_id')['input_assay_id'], on='regression_task_id')

T4_join.shape

T1 = T4_join[['input_compound_id', 'input_assay_id', 'standard_qualifier',
       'standard_value']]

print(T1.shape)
print(T1.columns)
T1.head()
